Final_Project_App/requirements.txt
streamlit==1.35.0
streamlit-webrtc==0.45.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
joblib==1.4.2
facenet-pytorch==2.5.3
scikit-learn==1.4.2
mediapipe==0.10.14
av==10.0.0
xgboost==2.0.3